预配组板页面的数据修改，修复分配解决的2个bug

+++ app/src/main/java/qx/app/freight/qxappfreight/activity/ReturnGoodsActivity.java
@@ -119,7 +119,7 @@ public class ReturnGoodsActivity extends BaseActivity implements MultiFunctionRe
                     rcInfosEntity.setWaybillCode(mMyAgentListBean.getWaybillCode());
                     rcInfosEntity.setCargoId(mMyAgentListBean.getCargoId());
                     rcInfosEntity.setNumber(mMyAgentListBean.getNumber());
-                    rcInfosEntity.setWeight(mMyAgentListBean.getWeight());
+                    rcInfosEntity.setWeight((int) mMyAgentListBean.getWeight());
                     rcInfosEntity.setVolume(mMyAgentListBean.getVolume());
                     rcInfosEntity.setPackagingType(mMyAgentListBean.getPackagingType());
                     rcInfosEntity.setScooterId(mMyAgentListBean.getScooterId());
